Rewrote the transform and codegen phases
There were a lot of issues with the old codegen stage, which had
evolved over time.  This rewrite created a new representation for
low level code which is straight forward for the codegen to work
with, and which doesn't rely on any of the IDs or previous work
of the earlier phases.  Most of the complexity is now in the
transform stage, including closure conversion which now works!
In the process, I also made methods be closures.

There are still some glaring omissions and failing tests.  Loops
have not yet been re-implemented, and recursion is still broken.