Fixed forloops and fixed issue with resolve
Forloops were fixed by desugaring them during the refine phase
and turning them into while loops.  They still directly reference
a list, but I want to change them to an iterator type, which
shouldn't be too tricky.

I also fixed the resolve issue which allowed the implementation of
an Option type in libcore, which can be used by an iterator to signal
the end, because nil doesn't work with boxed types =(
The fix makes a direct reference to the vtable of the referenced
type and looks up the static (or not static) method in the vtable.