Fixed issues with scope closure conv and names
Fixed detection of a variable inside of the functions scope, but
outside of the current scope.  This allows for match arms to have
their own scope (assists with pattern matching bindings), and also
allows us to pass in tscope to class methods during transform
which fixes the transformed names of methods to include the class
name.

This also partially implements pattern matching bindings, but it
doesn't work because of issues with the type checker