Added Enums!  First rough implementation
There are a number of issues with the current implementation.  The
enum size is hardcoded because we can't calculate at complie-time
what the size should be to fit every variant.  We need the target
architecture to know that, which would be a larger task to add.
It also does not allow nested types because they are implemented
as immediate structs and not pointers to structs.  The latter would
make it much easier to make nested types, but we'd have to allocate
the data for all of them (including option types).  A variant can
either have no arguments or a tuple of arguments.  This was due
to using the function call mechanics to implement enum constructors.
It's also very messy.  I'd like to clean things up, maybe by using
traits for some Def implementations, especially in the transform
stage