Fixed segfault caused by mismatched return types in try/catch
When a function was returning a non-unit value, and also possibly
throwing an exception, and there was a try/catch block that either
returned the value of the try body, or the value of the exception
handler arm which only printed, the return values would mismatch,
and the llvm phi instruction would cause a segfault, or sometimes
a "cannot emit physreg copy instruction" error if an integer was
returned.

This was caused because the typechecker was not correctly checking
the try block's return type with the exception arm's return type.
All the arms had to match, but the body could be different.  Now
a type error will occur during compiling such that you must make
the overall block's return type match