Fixed a few bugs in pattern matching
The enum pattern matching type checking was not using the type hint,
so when matching on an enum variant, who's type contains a type param,
it would use the variable and not the concrete type, even if one was
known from the type of the match condition.  That meant printing the
bound enum variant value would be ambiguous.

That uncovered another issue where type annotations in an enum pattern
wouldn't work, and it was instead being parsed as AST, which was preventing
a more helpful error message.  Pattern matching was extended to use
its own tuple and record parsing, which allows for type annotations to be
parsed correctly.