Fixed list test failure by using a hidden define in method invoke
The way method invocations added the object as the first argument
to function calls has been an issue for a while.  Previously the
ast was duplicated between the accessor and the first argument.
Now, a new definition will be created and the object ast assigned
to it.  The originial accessor will then reference the definition,
as will the first argument