Fixed and refactored the main and run functions
There was an issue with LLVM segfaulting when using getelementptr
to modify values within a closure tuple, but it turned out to be
related to the LLVM context.  A few functions weren't using the
InContext version and fixing that fixed those segfaults.  There is
still an issue with using imported functions within a method