Modified traits to use universals with constraints
Instead of a trait being used as Type::Object(), there are now
"where" clauses on functions, decls, and classes, which adds a trait
constraint to a universal variable in the declaration.  All universals
are now compiled as trait objects (although that will probably change)
and either have a NULL vtable or a vtable for the constraint trait.
It's now possible to call a method on a universal, which will attempt
to use the constraint to find a trait which has the requested method