Fixed regression with scope names, and a bug with overlapping traitdef funcs
The bug was caused when a traitdef had a function named "str" which is also
a function defined at the top level.  The traitdef was putting the new
binding into its vars scope, but traitimpl was putting it in the tscope
of the impl.  The implementation was calling the top level "str" function
but was instead getting linked to the trait impl method, but that error wasn't
caught and it was causing a segfault.  The new name-conflict contains the code
that caused this bug