Fixed matching of tuples, records, and enums
Previously the actual resulting boolean for each element in a record,
tuple, or enum variant arguments was being put into the expr list but
not never referenced.  It now joins the values together using LLExpr::Cmp
and also using an LLExpr::Phi for the enum variant number, which only if
true will then execute the comparisons for the arguments and return the
result.