Added explicit packing/unpacking of universal variables
Previously the actual boxing and unboxing of non-pointer data types
was done during the codegen stage by checking at every possible
location if the LLVM types don't align and if one is a pointer and
the other is not, but this is a bit messy.  In order to implement
traits, I needed to add checks throughout the transform stage to
see if any conversion occurred between a universal type and anything
else, in order to pack the object with the vtable for the trait
constraints on that value, which made things even more complicated.

To fix this, I added a stage after validation but before transform
which does the checks and modifies the AST to insert pack and unpack
AST nodes.  The transform stage could then just do the conversions
at the previously found locations.  This required me to add otherwise
unused AST node kinds, which isn't as clean as I'd like, but it's not
bad.  An alternative would be making an MIR and converting into MIR
and then doing the checks